requirements.txt
pandas>=1.3.0
numpy>=1.21.0
prophet>=1.1.0
scikit-learn>=1.0.0
hydra-core>=1.1.0
plotly>=5.0.0
openpyxl>=3.0.0